Add files via upload
Allows the caller to tell mslinks that the target is a directory. Useful when running in a context that has different file system access than the user will have. Such as a Tomcat server running as a service, that doesn't have access to the target of the link, so can't determine if the link is to a file or a directory.